+++ apps/client/pages/index.tsx
@@ -1,8 +1,12 @@
 import React from 'react';
-import { HomeLayout } from '@readable/home/feature-home';
+import { HomePage } from '@readable/home/feature-home';
+import { ComingSoonPage } from '@readable/home/feature-coming-soon';
+import { useAuth } from '@readable/shared/data-access-auth';
 
 function Index() {
-  return <HomeLayout />;
+  const { authenticated, logout } = useAuth();
+
+  return <div>{authenticated ? <HomePage /> : <ComingSoonPage />}</div>;
 }
 
 export default Index;
