--- libs/home/feature-home/src/lib/home/index.ts
@@ -1,1 +1,0 @@
-export { Home } from './home';
